+++ Model/Adminhtml/Source/Debug.php
@@ -17,8 +17,9 @@ class Debug implements ArrayInterface
     public function toOptionArray()
     {
         return [
-            2 => __('Split the log by days'),
-			1 => __('Save in single file'),
+            3 => __('Split the log by days'),
+			2 => __('Save in single file'),
+			1 => __('Save both files'),
 			0 => __('No'),
         ];
     }